*** empty log message ***

+++ c_ri/c_ri_get_B.m
@@ -92,6 +92,7 @@ FGMPATH = '/share/fgm_cal';
 if ~strcmp(h(1:end-1),'sanna.irfu.se')
 	FGMPATH = ['/net/sanna/export' FGMPATH];
 end
+if exist([FGMPATH '/tmp_att']), unix(['rm ' FGMPATH '/tmp_att']);end
 
 if nargout,  % return B
     to_file=tempname;
@@ -123,4 +124,4 @@ else
 end
 %removes the temporary files
 unix(['rm ' d_source]);
-unix(['rm ' FGMPATH '/tmp_att']);
+if exist([FGMPATH '/tmp_att']), unix(['rm ' FGMPATH '/tmp_att']);end
